refactor(renaming): split files.ts into focused modules
Extract single 427-line files.ts into four modules by responsibility:
- paths.ts: shared path helpers (validatePathTraversal, normalizePath, isSamePath)
- files.ts: single-file operations (getFileInfo, fileExists, renameFile, generateUniqueFilename, deduplicatePaths)
- batch.ts: batch orchestration and conflict detection (batchRename, checkConflicts)
- scan.ts: directory scanning (scanDirectoryForFiles, scanDirectoryRecursive)

Split monolithic files.test.ts into files.test.ts (17 tests) and
batch.test.ts (15 tests) mirroring the source split.

Also apply CodeRabbit review fixes: normalize realpath results for
case-insensitive comparison in paths.ts, use isSamePath for directory
rename child fixup in batch.ts.
